requirements.txt
albumentations==1.3.1
matplotlib==3.7.0
numpy==1.23.5
opencv_python_headless==4.8.0.74
Pillow==9.4.0
Pillow==10.0.0
segmentation_models_pytorch==0.3.3
streamlit==1.24.1
torch==1.12.1
torch==2.0.1
torchvision==0.15.2
